selphy/requirements.txt
activate==0.2.1
alembic==1.13.1
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
blinker==1.8.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configparser==7.0.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
Django==4.0.10
django-celery-beat==2.3.0
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-filter==23.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-modeltranslation==0.18.4
django-rosetta==0.9.9
django-timezone-field==6.0.1
django-utils-six==2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==19.3.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
idna==3.4
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.3
kombu==5.3.2
Mako==1.3.5
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.26.0
openapi-codec==1.3.2
packaging==24.1
Pillow==10.0.0
polib==1.2.0
prompt-toolkit==3.0.39
psycopg==3.1.18
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.0.3
requests==2.31.0
setuptools==74.0.0
simplejson==3.19.3
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==3.0.3
WTForms==3.1.2
